requirements.txt
aiohttp
python-dotenv
discord.py
flask